requirements.txt
anyio==4.4.0
appdirs==1.4.4
attrs==23.2.0
backoff==2.2.1
boilerpy3==1.0.7
cattrs==23.2.3
certifi==2024.6.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==7.0.0
docopt==0.6.2
elastic-transport==8.13.1
elasticsearch==8.14.0
etelemetry==0.3.1
Events==0.5
faiss-cpu==1.8.0.post1
farm-haystack==1.26.2
filelock==3.15.4
fitz==0.0.1.dev2
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
inflect==7.3.0
isodate==0.6.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lazy-imports==0.3.1
looseversion==1.3.0
lxml==5.2.2
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
num2words==0.5.13
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pillow==10.3.0
platformdirs==4.2.2
posthog==3.5.0
prompthub-py==4.0.0
prov==2.0.1
pydantic==1.10.17
pydot==2.0.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
quantulum3==0.9.2
rank-bm25==0.2.2
rdflib==6.3.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-cache==0.9.8
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
sentence-transformers==3.0.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sseclient-py==1.8.0
sympy==1.12.1
tenacity==8.4.2
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.4
traits==6.3.2
transformers==4.39.3
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.2
